Move MetadataSelector to new module with optional ipywidgets import
