+++ neurotic/gui/__init__.py
@@ -7,3 +7,5 @@ from ..gui.epochencoder import *
 from ..gui.icons import *
 from ..gui.config import *
 from ..gui.standalone import *
+
+from ..gui.notebook import *
